Stop computing error info in move path builder.
